// irmin-console/src/types/Wordpress.ts
export interface LogoCloudSection {
  acf_fc_layout: 'logo_cloud';
  title: string;
  logos: Logo[];
}

interface Logo {
  logo: string | number;
  title: string;
}